:sparkles: Enable to run `deletePage()` with an existing socket
